+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Changed
+
+- Updated checksums of known-types [#667](https://github.com/polkadot-api/polkadot-api/pull/667)
+- Remove unused dependencies
+
 ## 0.7.6 - 2024-08-26
 
 ### Fixed